requirements.txt
defusedxml==0.7.1
passlib==1.7.4
vobject>=0.9.6
python-dateutil==2.8.2
redis==4.3.4